backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.29.0
pydantic==2.10.3
pydantic-settings==2.7.0
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.4.2
shap==0.45.1
fairlearn==0.10.0
google-cloud-aiplatform==1.74.0
google-cloud-storage==2.14.0
google-cloud-firestore==2.14.0
python-multipart==0.0.19
aiofiles==24.1.0
httpx==0.28.0
pytest